requirements.txt
asgiref==3.3.1
Django==3.1.5
django-countries==7.0
django-internationalflavor==0.4.1
Pillow==8.1.0
python-decouple==3.4
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
